+++ source/BaseEngine/GUI/GUIText.cpp
@@ -23,7 +23,7 @@ void CGUIText::DrawText(const FontShader & shader, const CFont& font) const
 	glColor4fv(active_color);
 	shader.loadTranslation(m_Position);
 	glScalef(m_FontSize, m_FontSize, m_FontSize);
-	font.Print(m_Position.x, m_Position.y, m_Text.c_str());
+	font.Print(static_cast<const int>(m_Position.x), static_cast<const int>(m_Position.y), m_Text.c_str());
 	glPopMatrix();
 	shader.Stop();
 }
